# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "All APIs related to Karchu.",
        "title": "Karchu API",
        "contact": {},
        "version": "1.0"
    },
    "host": "karchu.onrender.com",
    "basePath": "/",
    "paths": {
...
        "/v1/net-amount/categories": {
            "post": {
                "description": "get money spent on each category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Net-Amount"
                ],
                "summary": "get money spent on each category",
                "parameters": [
                    {
                        "description": "enter Email,Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.NetAmountByCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "money spent on each category as list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/transactions": {
            "post": {
                "description": "create a transaction with category V2",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions",
                    "V2"
                ],
                "summary": "creates a transaction for a user V2",
                "parameters": [
                    {
                        "description": "enter Email,Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateTransactionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.SuccessRes"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "delete a transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "delete transaction for a user",
                "parameters": [
                    {
                        "description": "enter Email,Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.DeleteTransactionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
        "/v1/transactions/all": {
            "post": {
                "description": "returns transactions as string for UI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions",
                    "V2"
                ],
                "summary": "Get transactions list as string of user",
                "parameters": [
                    {
                        "description": "enter Email,Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.GetTransactionsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns transaction strings as list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/transactions/get": {
            "post": {
                "description": "returns transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions",
                    "V2"
                ],
                "summary": "Get transactions of user",
                "parameters": [
                    {
                        "description": "enter Email,Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateTransactionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/transactions/last-n": {
            "post": {
                "description": "Get last N transaction list of user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get last N transactions of user",
                "parameters": [
                    {
                        "description": "enter Email,Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.GetLastNTransactionsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "last N transactions list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/transactions/str": {
            "delete": {
                "description": "delete a transaction for a given trans string",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "delete transaction for a user",
                "parameters": [
                    {
                        "description": "enter Email,Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.DeleteTransactionFromTransStringReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
        "/v1/user/": {
            "post": {
                "description": "creates a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "create a user",
                "parameters": [
                    {
                        "description": "enter Email, Name and Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
        "/v1/user/auth": {
            "post": {
                "description": "returns userId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Authorizes the user creds",
                "parameters": [
                    {
                        "description": "enter Email and Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UserId",
                        "schema": {
                            "type": "int"
                        }
                    }
                }
            }
        },
...
        "/v2/trans-split-with-one": {
            "post": {
                "description": "create a transaction and also split",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trans-split-with-one"
                ],
                "summary": "creates a transaction and also split",
                "parameters": [
                    {
                        "description": "add transaction and also split",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.TransactionAndSplitWithOneReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.SuccessRes"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "requests.CreateTransactionReq": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "splitTag": {
                    "type": "string"
                }
            }
        },
        "requests.CreateUserReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "requests.DeleteTransactionFromTransStringReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "transString": {
                    "type": "string"
                }
            }
        },
        "requests.DeleteTransactionReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "transactionId": {
                    "type": "integer"
                }
            }
        },
...
        "requests.GetLastNTransactionsReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "lastN": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "requests.GetTransactionsReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "requests.NetAmountByCategoryReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "requests.TransactionAndSplitWithOneReq": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "friendName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "splitAmount": {
                    "type": "integer"
                },
                "splitTag": {
                    "type": "string"
                }
            }
        },
...
        "requests.UserReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "responses.SuccessRes": {
            "type": "object",
            "properties": {
                "success_code": {
                    "type": "string"
                },
                "success_message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}